// src/types/rbac.ts
import { RoleCode } from './auth';

export type PermissionCode =
  // Members & KYC
  | 'MEMBER:PROFILE:VIEW:ALL'
  | 'MEMBER:PROFILE:VIEW:OWN'
  | 'MEMBER:PROFILE:CREATE:ALL'
  | 'MEMBER:PROFILE:CREATE:SELF'
  | 'MEMBER:PROFILE:UPDATE:ALL'
  | 'MEMBER:PROFILE:UPDATE:OWN'
  | 'MEMBER:PROFILE:DELETE:REQUEST'
  | 'MEMBER:PROFILE:DELETE:APPROVE'
  | 'MEMBER:PROFILE:RESTORE'
  | 'MEMBER:PROFILE:SUSPEND'
  | 'MEMBER:PROFILE:ACTIVATE'
  // Receipts & Verification
  | 'MEMBER:RECEIPT:UPLOAD:OWN'
  | 'MEMBER:RECEIPT:VIEW:ALL'
  | 'MEMBER:RECEIPT:VERIFY'
  | 'MEMBER:RECEIPT:APPROVE'
  | 'MEMBER:RECEIPT:REJECT'
  // Savings
  | 'SAVING:ACCOUNT:VIEW:ALL'
  | 'SAVING:ACCOUNT:VIEW:OWN'
  | 'SAVING:ACCOUNT:CREATE'
  | 'SAVING:ACCOUNT:CLOSE'
  | 'SAVING:INTEREST:EXECUTE'
  | 'SAVING:INTEREST:APPROVE'
  | 'SAVING:TIMEDEPOSIT:OVERRIDE'
  // Shares & Equity
  | 'SHARE:ACCOUNT:VIEW:ALL'
  | 'SHARE:ACCOUNT:VIEW:OWN'
  | 'SHARE:PURCHASE:INITIATE:OWN'
  | 'SHARE:PURCHASE:PROCESS'
  | 'SHARE:CONVERT:INITIATE:OWN'
  | 'SHARE:CONVERT:PROCESS'
  // Loans & Credit
  | 'LOAN:APPLICATION:CREATE:OWN'
  | 'LOAN:APPLICATION:CREATE'
  | 'LOAN:APPLICATION:VIEW:ALL'
  | 'LOAN:APPLICATION:VIEW:OWN'
  | 'LOAN:APPLICATION:VERIFY'
  | 'LOAN:APPLICATION:APPROVE'
  | 'LOAN:APPLICATION:REJECT'
  | 'LOAN:DISBURSEMENT:EXECUTE'
  | 'LOAN:REPAYMENT:RECORD'
  | 'LOAN:PENALTY:WAIVE'
  | 'LOAN:SCHEDULE:PRINT'
  // Transactions & Ledgers
  | 'TXN:DEPOSIT:CREATE'
  | 'TXN:WITHDRAWAL:CREATE'
  | 'TXN:WITHDRAWAL:APPROVE:LARGE'
  | 'TXN:REVERSAL:REQUEST'
  | 'TXN:REVERSAL:APPROVE'
  | 'TXN:LEDGER:VIEW:ALL'
  | 'TXN:PASSBOOK:VIEW:OWN'
  // Accounting & GL
  | 'ACCOUNTING:COA:MANAGE'
  | 'ACCOUNTING:JOURNAL:CREATE'
  | 'ACCOUNTING:JOURNAL:POST'
  | 'ACCOUNTING:STATEMENTS:GEN'
  | 'ACCOUNTING:AUDIT:VIEW:ALL'
  // Reports
  | 'REPORT:OPERATIONAL:EXPORT'
  | 'REPORT:FINANCIAL:EXPORT'
  | 'REPORT:DEFAULTERS:VIEW'
  | 'REPORT:PREDICTIONS:VIEW'
  // Support & Comms
  | 'SUPPORT:TICKET:CREATE:OWN'
  | 'SUPPORT:TICKET:MANAGE:ALL'
  | 'SUPPORT:TICKET:VIEW:OWN'
  | 'NOTIF:DISPATCH:MANUAL'
  | 'NOTIF:HISTORY:VIEW:ALL'
  // System & Security Governance
  | 'SYSTEM:USER:MANAGE'
  | 'SYSTEM:ROLE:MANAGE'
  | 'SYSTEM:SETTINGS:UPDATE'
  | 'SYSTEM:AUDIT:VIEW:ALL';

export interface RoleDefinition {
  code: RoleCode;
  name: string;
  description: string;
  portalPrefix: string;
  isSystem: boolean;
  permissions: PermissionCode[];
}